Project 3
For this project we will use an Arduino Uno or equivalent controller board and components of choice.

Standards describing communication in controller networks often specify a device state machine that all slave nodes in the network must implement, in order to provide a common interface to the network master.

In the CANopen standard used in industrial automation this is the Network Management (NMT) slave state machine (Links to an external site.).

NMT slave state machineThis state machine has four main states that control the input-output behavior, LED indication, and which commands are accepted and acted on.

In the Initialization state, the device starts up and initializes its internal parameters.  Following initialization, transition to Pre-operational is autonomous.
In the Pre-operational (or configuration) state the LED should blink at 1 Hz and the device should accept configuration commands, a set operational command that triggers transition into the Operational state, and a reset command to trigger transition to the Initialization state.
In the Operational state, the normal input-output behavior should be active.  If the LED is not used as an output device, then the LED should be continuously on.  The device should accept a set pre-operational command that triggers transition to the Pre-operational state, and a reset command to trigger transition to the Initialization state.
In the Stopped state the LED should blink at 2 Hz.  The device should accept a clear fault command that sends it back from the Stopped state.
In this project you will implement the state machine for a simple device with one analog input controlling one analog output.  You can choose a sensor and output of your liking (e.g. a photo-resistor controlling LED intensity).

The project is in three parts.  For each part, design and draw the UML state diagram for your system using PlantUML.  The diagram should show the events that trigger state transitions and the resulting activity (e.g. entry/exit) triggered on state transition.

Implement the state machine using the state behavior pattern.  An example pattern is given.

Part 1
Implement the Initialization and Operational states and the reset command.  Include any initialization that your sensor may require.  You can use the Arduino Serial class to issue simple commands from the keyboard.

Part 2
Add the Pre-operational state and a configuration parameter that visually affects the Operational behavior.

Add serial commands for configuration and commands to  transition between the Pre-operational and Operational states.

Part 3
Add fault detection that autonomously triggers transition  into the stopped state.  This can be a type of electrical failure (e.g. open or short circuit).

Make a short video showing the sensor controlled blinking in action.  Upload the video to YouTube and submit a link.

Upload the code to a git repository.

This is a group assignment.